Fixed Issue #14
- Export To CNTK now exports to the furthest visted frame instead of the last frame in a video

- Fixed bug where metadata from previous exports was persisting

- Made the videotagging var higher scope to minimize duplicate code

-  Added getUnlabeledRegions(FrameId) to videotagging control

+++ index.js
@@ -3,7 +3,16 @@ var basepath = remote.app.getAppPath();
 var dialog = remote.require('electron').dialog;
 var path = require('path');
 var fs = require('fs');
+
+var furthestVisitedFrame; //keep track of the furthest visited frame
+var videotagging;
+
 const ipcRenderer = require('electron').ipcRenderer;
+
+function updateFurthestVisitedFrame(){
+    var currentFrame = videotagging.getCurrentFrame();
+    if (furthestVisitedFrame < currentFrame) furthestVisitedFrame = currentFrame;            
+}
 
 function fileSelcted() {
 
@@ -20,10 +29,8 @@ function fileSelcted() {
       document.getElementById('load-message').style.display = "none";
       document.getElementById('load-form-container').style.display = "block";
 
-      //set title indicator
-       $('title').text(`Video Tagging Job Configuration: ${path.basename(fileName[0], path.extname(fileName[0]))}`);
-
-      $('#inputtags').tagsinput('removeAll');
+      $('title').text(`Video Tagging Job Configuration: ${path.basename(fileName[0], path.extname(fileName[0]))}`);      //set title indicator
+      $('#inputtags').tagsinput('removeAll');//remove all previous tag labels
 
       try{
         config = require(`${fileName}.json`);
@@ -40,8 +47,7 @@ function fileSelcted() {
       }
       document.getElementById('loadButton').addEventListener('click', function (e) {
 
-        var videotagging = document.getElementById('video-tagging');
-
+        videotagging = document.getElementById('video-tagging');
         videotagging.framerate = document.getElementById('framerate').value;
         videotagging.regiontype = document.getElementById('regiontype').value;
         videotagging.multiregions = document.getElementById('MultiRegions').checked ? "1":"0";
@@ -65,6 +71,11 @@ function fileSelcted() {
         //set title indicator
         $('title').text(`Video Tagging Job: ${path.basename(fileName[0], path.extname(fileName[0]))}`);
 
+        //track furthestVisitedFrame
+        furthestVisitedFrame = 1;
+        videotagging.video.removeEventListener("timeupdate", updateFurthestVisitedFrame); //remove old listener
+        videotagging.video.addEventListener("timeupdate",updateFurthestVisitedFrame);
+
       });
     }
     else {
@@ -75,7 +86,6 @@ function fileSelcted() {
 }
 
 function save() {
-    var videotagging = document.getElementById('video-tagging');
     var saveObject = {
       "frames" : videotagging.frames,
       "inputTags": document.getElementById('inputtags').value,
@@ -90,8 +100,6 @@ function save() {
 function exportCNTK() {
 
   $("<div class=\"loader\"></div>").appendTo($("#videoWrapper"));
-
-  var videotagging = document.getElementById('video-tagging');
 
   //make sure paths exist
   if (!fs.existsSync(`${basepath}/cntk`)) fs.mkdirSync(`${basepath}/cntk`);
@@ -108,36 +116,51 @@ function exportCNTK() {
   var canvasContext = frameCanvas.getContext("2d");
 
   // start exporting frames using the timeupdate eventListener
+  videotagging.video.removeEventListener("timeupdate", updateFurthestVisitedFrame); //stop recording frame movment
   videotagging.video.addEventListener("timeupdate", saveFrames);
   videotagging.video.currentTime = 0;
   videotagging.playingCallback();
 
   function saveFrames(){
 
+    var frameId = videotagging.getCurrentFrame();
     //if last frame removeEventListener and loader
-    if (videotagging.video.currentTime >= videotagging.video.duration){
+    if (( frameId >= furthestVisitedFrame) ) { //|| (videotagging.video.currentTime >= videotagging.video.duration)
       videotagging.video.removeEventListener("timeupdate", saveFrames);
+      videotagging.video.addEventListener("timeupdate", updateFurthestVisitedFrame);
       $(".loader").remove();
     }
 
-    var frameId = videotagging.frameText.innerText;
-
     //set default writepath to the negative folder
-    var writePath = `${framesPath}/negative/${path.basename(videotagging.src[0], path.extname(videotagging.src[0]))}_frame_${frameId}.jpg`;
-
+    var writePath = `${framesPath}/negative/${path.basename(videotagging.src[0], path.extname(videotagging.src[0]))}_frame_${frameId}.jpg`; //defaults to negative
+    var positiveWritePath = `${framesPath}/positive/${path.basename(videotagging.src[0], path.extname(videotagging.src[0]))}_frame_${frameId}.jpg`;
     //If frame contains tags generate the metadata and save it in the positive directory
-    if (videotagging.frames.hasOwnProperty(frameId)){
+    var frameIsTagged = videotagging.frames.hasOwnProperty(frameId) && (videotagging.frames[frameId].length > 0);
+    var frameContainsLabels = (videotagging.getUnlabeledRegionTags(frameId).length != videotagging.frames[frameId].length);
+    if (frameIsTagged && frameContainsLabels){
+        //clear metadata if image exists from last run
+        if(fs.existsSync(writePath))fs.unlinkSync(writePath);
+        if(fs.existsSync(positiveWritePath)){
+          fs.unlinkSync(positiveWritePath.replace('.jpg', '.bboxes.labels.tsv'));
+          fs.unlinkSync(positiveWritePath.replace('.jpg', '.bboxes.tsv'));
+        }
+        //genereate metadata from tags
         videotagging.frames[frameId].map(function(tag){
               if (!tag.tags[tag.tags.length-1]) {
-                console.log(`frame ${frameId} region ${tag.name} has no label`);
-                return;
+                return console.log(`frame ${frameId} region ${tag.name} has no label`);
               }
-              writePath = `${framesPath}/positive/${path.basename(videotagging.src[0], path.extname(videotagging.src[0]))}_frame_${frameId}.jpg`;
               var stanW = videotagging.video.videoWidth/tag.width;
               var stanH = videotagging.video.videoHeight/tag.height;
-              fs.appendFile(writePath.replace('.jpg', '.bboxes.labels.tsv'), `${tag.tags[tag.tags.length-1]}\n`, function (err) {});
-              fs.appendFile(writePath.replace('.jpg', '.bboxes.tsv'), `${parseInt(tag.x1 * stanW)}\t${parseInt(tag.y1 * stanH)}\t${parseInt(tag.x2 * stanW)}\t${parseInt(tag.y2 * stanH)}\n`, function (err) {});
+              fs.appendFile(positiveWritePath.replace('.jpg', '.bboxes.labels.tsv'), `${tag.tags[tag.tags.length-1]}\n`, function (err) {});
+              fs.appendFile(positiveWritePath.replace('.jpg', '.bboxes.tsv'), `${parseInt(tag.x1 * stanW)}\t${parseInt(tag.y1 * stanH)}\t${parseInt(tag.x2 * stanW)}\t${parseInt(tag.y2 * stanH)}\n`, function (err) {});
         });
+
+        writePath = positiveWritePath; // set write path to positve write path
+    }
+    else if(fs.existsSync(positiveWritePath)){ //tags have been removed clear positive data if itexists from last run
+        fs.unlinkSync(positiveWritePath);
+        fs.unlinkSync(positiveWritePath.replace('.jpg', '.bboxes.tsv'));
+        fs.unlinkSync(positiveWritePath.replace('.jpg', '.bboxes.labels.tsv'));
     }
 
     //draw the frame to the canvas
@@ -147,9 +170,9 @@ function exportCNTK() {
 
     //write canvas to file and change frame
     console.log('saving file', writePath);
-    fs.writeFileSync(writePath, buf);
-    videotagging.stepFwdClicked();
+    if(!fs.existsSync(writePath))fs.writeFileSync(writePath, buf);
+    if (frameId < furthestVisitedFrame) videotagging.stepFwdClicked();
 
   }
-  
+
 }
